Change recount_threshold from a DecimalField (optimal) to a FloatField (non-optimal) to work around a MS-SQL issue. Old code left in as a comment if I have some troubleshooting time.
git-svn-id: https://punch.cs.sfu.ca/svn/CMPT275-1121-g-group2/trunk@623 a35e499e-f4d2-495b-8bd3-10c317ad12fd

+++ election/models.py
@@ -17,8 +17,13 @@ class Election(models.Model):
     status = models.CharField(max_length=3, choices=STATUS_CHOICES, default='BEF')
     description = models.CharField(max_length=128, verbose_name="Name", help_text="For later identification")
     start = models.DateField(verbose_name="Date", help_text="Start date of the election.  Format: YYYY-MM-DD")
-    recount_threshold = models.DecimalField(max_digits=4, decimal_places=2,
-        verbose_name="Recount Threshold",
+    # recount_threshold = models.DecimalField(max_digits=4, decimal_places=2,
+        # verbose_name="Recount Threshold",
+        # help_text="Percentage (0.00% to 99.99%) of difference to initiate a recount atuomatically.",
+        # validators=[MinValueValidator(0.00),MaxValueValidator(99.99)],
+        # default=0.01)
+        
+    recount_threshold = models.FloatField(verbose_name="Recount Threshold",
         help_text="Percentage (0.00% to 99.99%) of difference to initiate a recount atuomatically.",
         validators=[MinValueValidator(0.00),MaxValueValidator(99.99)],
         default=0.01)
